requirements.txt
jsonschema>=4.20.0
PyMuPDF>=1.24.0
paddleocr>=2.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=8.0.0

